+++ lib/src/client/connection.dart
@@ -43,7 +43,8 @@ abstract class ClientConnection {
 
   /// Start a request for [path] with [metadata].
   GrpcTransportStream makeRequest(String path, Duration timeout,
-      Map<String, String> metadata, ErrorHandler onRequestFailure);
+      Map<String, String> metadata, ErrorHandler onRequestFailure,
+      {CallOptions callOptions});
 
   /// Shuts down this connection.
   ///